+++ packages/client/src/adapters/fetch/rpc-link.test.ts
@@ -1,7 +1,8 @@
 import { ORPCError } from '@orpc/contract'
 import { os } from '@orpc/server'
 import { RPCHandler } from '@orpc/server/fetch'
-import { getEventMeta, isAsyncIteratorObject, withEventMeta } from '@orpc/standard-server'
+import { isAsyncIteratorObject } from '@orpc/shared'
+import { getEventMeta, withEventMeta } from '@orpc/standard-server'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { supportedDataTypes } from '../../../tests/shared'
 import { RPCLink } from './rpc-link'
