+++ packages/client/src/adapters/fetch/rpc-link.ts
@@ -1,8 +1,8 @@
 import type { Value } from '@orpc/shared'
+import type { StandardBody } from '@orpc/standard-server'
 import type { ClientContext, ClientLink, ClientOptionsOut } from '../../types'
 import type { FetchWithContext } from './types'
-import { trim, value } from '@orpc/shared'
-import { isAsyncIteratorObject, type StandardBody } from '@orpc/standard-server'
+import { isAsyncIteratorObject, trim, value } from '@orpc/shared'
 import { toFetchBody, toStandardBody } from '@orpc/standard-server-fetch'
 import { ORPCError } from '../../error'
 import { createAutoRetryEventIterator, type EventIteratorReconnectOptions } from '../../event-iterator'
