+++ packages/client/src/rpc/serializer.test.ts
@@ -1,5 +1,6 @@
 import { ORPCError } from '@orpc/contract'
-import { ErrorEvent, getEventMeta, isAsyncIteratorObject, parseEmptyableJSON, withEventMeta } from '@orpc/standard-server'
+import { isAsyncIteratorObject, parseEmptyableJSON } from '@orpc/shared'
+import { ErrorEvent, getEventMeta, withEventMeta } from '@orpc/standard-server'
 import { supportedDataTypes } from '../../tests/shared'
 import { RPCSerializer } from './serializer'
