+++ packages/server/src/adapters/standard/rpc-codec.ts
@@ -1,8 +1,9 @@
 import type { ORPCError } from '@orpc/client'
+import type { StandardBody, StandardRequest, StandardResponse } from '@orpc/standard-server'
 import type { AnyProcedure } from '../../procedure'
 import type { StandardCodec, StandardParams } from './types'
 import { RPCSerializer } from '@orpc/client/rpc'
-import { parseEmptyableJSON, type StandardBody, type StandardRequest, type StandardResponse } from '@orpc/standard-server'
+import { parseEmptyableJSON } from '@orpc/shared'
 
 export interface StandardCodecOptions {
   serializer?: RPCSerializer
