+++ packages/shared/src/index.ts
@@ -2,6 +2,8 @@ export * from './chain'
 export * from './error'
 export * from './function'
 export * from './interceptor'
+export * from './iterator'
+export * from './json'
 export * from './object'
 export * from './types'
 export * from './value'
