+++ packages/shared/src/iterator.ts
@@ -1,0 +1,7 @@
+export function isAsyncIteratorObject(maybe: unknown): maybe is AsyncIteratorObject<any, any, any> {
+  if (!maybe || typeof maybe !== 'object') {
+    return false
+  }
+
+  return Symbol.asyncIterator in maybe && typeof maybe[Symbol.asyncIterator] === 'function'
+}
