+++ packages/shared/src/json.test.ts
@@ -1,0 +1,7 @@
+import { parseEmptyableJSON } from './json'
+
+it('parseEmptyableJSON', () => {
+  expect(parseEmptyableJSON('')).toBeUndefined()
+  expect(parseEmptyableJSON('{}')).toEqual({})
+  expect(parseEmptyableJSON('{"foo":"bar"}')).toEqual({ foo: 'bar' })
+})
