+++ packages/shared/src/json.ts
@@ -1,0 +1,7 @@
+export function parseEmptyableJSON(text: string): unknown {
+  if (!text) {
+    return undefined
+  }
+
+  return JSON.parse(text)
+}
