+++ packages/standard-server-fetch/playground/event-source.ts
@@ -1,4 +1,4 @@
-import { isAsyncIteratorObject } from '@orpc/standard-server'
+import { isAsyncIteratorObject } from '@orpc/shared'
 import { toStandardRequest } from '../src/request'
 import { toFetchResponse } from '../src/response'
 import { serve } from '@hono/node-server'
