+++ packages/standard-server-fetch/src/event-source.test.ts
@@ -1,4 +1,5 @@
-import { ErrorEvent, getEventMeta, isAsyncIteratorObject, UnknownEvent, withEventMeta } from '@orpc/standard-server'
+import { isAsyncIteratorObject } from '@orpc/shared'
+import { ErrorEvent, getEventMeta, UnknownEvent, withEventMeta } from '@orpc/standard-server'
 import { toEventIterator, toEventStream } from './event-source'
 
 describe('toEventIterator', () => {
