+++ packages/standard-server-node/package.json
@@ -41,6 +41,7 @@
     "type:check": "tsc -b"
   },
   "dependencies": {
+    "@orpc/shared": "workspace:*",
     "@orpc/standard-server": "workspace:*"
   },
   "devDependencies": {
