+++ packages/standard-server-node/src/body.ts
@@ -2,7 +2,8 @@ import type { StandardBody, StandardHeaders } from '@orpc/standard-server'
 import type { Buffer } from 'node:buffer'
 import type { NodeHttpRequest } from './types'
 import { Readable } from 'node:stream'
-import { contentDisposition, isAsyncIteratorObject, parseContentDisposition, parseEmptyableJSON } from '@orpc/standard-server'
+import { isAsyncIteratorObject, parseEmptyableJSON } from '@orpc/shared'
+import { contentDisposition, parseContentDisposition } from '@orpc/standard-server'
 import { toEventIterator, toEventStream } from './event-source'
 
 export async function toStandardBody(req: NodeHttpRequest): Promise<StandardBody> {
