+++ packages/standard-server-node/src/event-source.ts
@@ -1,11 +1,11 @@
 import { Readable } from 'node:stream'
+import { parseEmptyableJSON } from '@orpc/shared'
 import {
   encodeEventMessage,
   ErrorEvent,
   EventDecoderStream,
   getEventMeta,
   isEventMetaContainer,
-  parseEmptyableJSON,
   UnknownEvent,
   withEventMeta,
 } from '@orpc/standard-server'
