+++ packages/standard-server-node/src/request.ts
@@ -1,6 +1,6 @@
 import type { StandardRequest } from '@orpc/standard-server'
 import type { NodeHttpRequest, NodeHttpResponse } from './types'
-import { once } from '@orpc/standard-server'
+import { once } from '@orpc/shared'
 import { toStandardBody } from './body'
 import { toAbortSignal } from './signal'
